# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Courtpit API",
    "description": "Courtpit REST API.\n\n**Community.** Every request is scoped to one community. Native clients send `X-Courtpit-Community: <slug>`; browsers are resolved from the host (`<slug>.courtpit.app` or a registered custom domain). The header wins over the host. A request without a resolvable community answers `400`, an unknown one `404`.\n\n**Authentication.** A session is an opaque token, sent as `Authorization: Bearer <token>` (native) or as the `courtpit_session` cookie (web; ask for it at sign-in with `X-Courtpit-Client: web`). Either one authenticates.\n\n**Errors.** Every error is `{ \"error\": { \"code\", \"message\" } }`; clients switch on `code`.\n\n**Lists** are cursor-paginated: `?cursor=&limit=` in, `{ items, next_cursor }` out.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/admin/leagues": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Creates a league in `draft` (admin). Publish it to open registration.",
        "operationId": "create_league",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeague"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/leagues/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Edits a league while it is still a draft (admin).",
        "operationId": "patch_league",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchLeague"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/leagues/{id}/cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Cancels a league that hasn't finished (admin). Its unplayed matches are cancelled too.",
        "operationId": "cancel_league",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/admin/leagues/{id}/finish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Finishes an active league whose `ends_at` has passed (admin). Without `force`, refuses\n(409 `unresolved_matches`) while any match is still reported or disputed; before the end\ndate it refuses with 409 `season_not_over`. Unplayed matches are cancelled and the final\nstandings, season points and promotion/relegation are recorded, as the lifecycle job does.",
        "operationId": "finish_league",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "Finish even though matches are still reported or disputed; they are left out of the\nfinal table.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "What went wrong."
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "description": "Machine-readable code plus a human-readable message.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code; clients switch on it. One of `bad_request` (400),\n`validation_failed` (422), `unauthorized` and `invalid_credentials` (401), `forbidden`\n(403), `not_found` (404), `conflict` (409), `rate_limited` (429) or `internal` (500)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation; never contains internal details."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token (native clients)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "courtpit_session",
        "description": "Session cookie (web clients); alternative to the bearer token."
      }
    }
  },
...
}